+++ apps/website-new/module-federation.config.ts
@@ -6,9 +6,10 @@ const LANGUAGES = ['zh', 'en'];
 
 const exposes = {
   // basic
-  [`./rspack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/rspack.mdx`,
-  [`./webpack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/webpack.mdx`,
-  [`./rspress-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/rspress.mdx`,
+  [`./plugins-overview-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/overview.mdx`,
+  [`./rspack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/rspack.mdx`,
+  [`./webpack-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/webpack.mdx`,
+  [`./rspress-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/plugins/rspress.mdx`,
   [`./cli-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/cli.mdx`,
   [`./type-prompt-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/type-prompt.mdx`,
   [`./css-isolate-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/basic/css-isolate.mdx`,
@@ -41,6 +42,11 @@ const exposes = {
   [`./configure-dts-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/dts.mdx`,
   [`./configure-dev-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/dev.mdx`,
   [`./configure-manifest-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/manifest.mdx`,
+  [`./configure-tree-shaking-dir-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/treeShakingDir.mdx`,
+  [`./configure-tree-shaking-shared-exclude-plugins-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/treeShakingSharedExcludePlugins.mdx`,
+  [`./configure-tree-shaking-shared-plugins-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/treeShakingSharedPlugins.mdx`,
+  [`./configure-inject-tree-shaking-used-exports-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/injectTreeShakingUsedExports.mdx`,
+
   // [`./configure-shareStrategy-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/shareStrategy.mdx`,
   // [`./configure-experiments-${LANGUAGE}`]: `./docs/${LANGUAGE}/configure/experiments.mdx`,
 
@@ -58,6 +64,10 @@ const exposes = {
 
   // blog
   [`./error-load-remote-${LANGUAGE}`]: `./docs/${LANGUAGE}/blog/error-load-remote.mdx`,
+
+  // performance
+  [`./shared-tree-shaking-overview-${LANGUAGE}`]: `./docs/${LANGUAGE}/guide/performance/shared-treeshaking.mdx`,
+  [`./node-${LANGUAGE}`]: `./docs/${LANGUAGE}/blog/node.mdx`,
 };
 
 export default createModuleFederationConfig({