+++ packages/chrome-devtools/src/index.tsx
@@ -1,5 +1,4 @@
 import { createRoot } from 'react-dom/client';
-
 import App from './App';
 
 (async () => {