+++ packages/modernjs/src/types/index.ts
@@ -1,6 +1,4 @@
 import { moduleFederationPlugin } from '@module-federation/sdk';
-import type { ModuleFederationPlugin as WebpackModuleFederationPlugin } from '@module-federation/enhanced';
-import type { ModuleFederationPlugin as RspackModuleFederationPlugin } from '@module-federation/enhanced/rspack';
 import type { StatsAssetResource } from '@module-federation/rsbuild-plugin/utils';
 
 export interface PluginOptions {
@@ -13,6 +11,7 @@ export interface PluginOptions {
     | boolean;
   remoteIpStrategy?: 'ipv4' | 'inherit';
   fetchServerQuery?: Record<string, unknown>;
+  secondarySharedTreeShaking?: boolean;
 }
 
 export type AssetFileNames = {
@@ -37,6 +36,7 @@ export interface InternalModernPluginOptions {
   remoteIpStrategy?: 'ipv4' | 'inherit';
   userConfig?: PluginOptions;
   fetchServerQuery?: Record<string, unknown>;
+  secondarySharedTreeShaking?: boolean;
 }
 
 export type BundlerPlugin = any;